DAL, DTO, DAO y POM. No inicia

+++ src/main/java/ejemploWeb2/aplicacion/dal/AlumnoRepository.java
@@ -11,6 +11,6 @@ import org.springframework.stereotype.Repository;
  *
  */
 @Repository
-public interface AlumnoRepository extends CrudRepository<Alumno,Integer> {
+public interface AlumnoRepository extends CrudRepository<Alumno,Long> {
 
 }
